Drop unused ParseError variants
An earlier design was going to try to detect when the "mul" keyword was
missing but other text was present, and emit an error matching this.
Likewise for the ::ExtraSpace variant. I was going to detect and fail on
the extra spaces. Instead, the parser was built to simply not recognize
these things and emit a more narrow message -- "couldn't find a number"
or "couldn't find the closing parenthesis", etc.