New rule checker unit tests
It turns out that Rust ranges *must* be `a > b` or they are empty. The
reversed range for the scan-left doesn't actually work because of this.

I've updated the unit tests to detect this case. There are forward and
backward scanning cases, and ones starting from the ends and not-the-
ends of the list. Time to fix.