Fix `check_rule` function -- double bug!
The previously found problem was that the Range syntax doesn't go in
reverse. This was fixed by just... copying the for loops. I can't use
an iterator and reversed iterator because they are different types.

I found another bug in the range bounds. For rightward iteration, I want
(idx..row.len()). But for leftward iteration, I need to go from idx down
to 0. Both were iterating to `row.len()` instead.