New test case for *another* bug!
When selecting active rules, I'm generating combinations of the input
sequence numbers. For valid sequences, all the applicable rules are
found correctly. For invalid sequences, the combinations will be flipped
and the intersection will drop them. The resulting set is passed forward
to the checker and the check passes... because I've eliminated the rules
that would be broken!

I've added a new active-rule selection test case that uses such a rule-
breaking sequence. Time to fix the bug.