Fix active rule selection bug! use `permutations`
The problem is that the combinations produced an ordered pair of numbers
based on the input sequence. When the input was already in-order
according to the primary ruleset, the resulting combos would be matching
Rules. When the input was *not* in order, the results would be reversed.
These reversed rules are not matched with the primary set and get
removed from the active set when computing the intersection.

If I use `permutations()` instead of `combinations()`, I get the numbers
in both directions. It's a lot more to discard, but I match the intended
rule in the mis-ordered sequences.